requirements.txt
# Video Analysis with Qwen3-VL Model via Ollama
# Requirements for video analysis using Ollama HTTP API

opencv-python>=4.8.0
pillow>=9.0.0
numpy>=1.21.0
tqdm>=4.64.0
requests>=2.28.0